fix: Update tag.yaml workflow to use GitHub Script for creating tags
The tag.yaml workflow has been updated to use the GitHub Script action for creating tags instead of using the `git tag` and `git push` commands. This change ensures that the correct variable name is used and that tags are pushed to the remote repository. The `createRef` function from the GitHub Script action is now used to create the tag reference with the appropriate parameters.